Replace LruCache implementation for layoutlib. do not merge.
The android version depends on a custom version of LinkedHashMap
which is not present on desktop VMs. This new implementation is done
in a way that minimizes the difference between the two.

Also some minor fixes.

(cherry picked from commit 3fa57f96b12a2148656cd9f38843df70f3ec6aca)

Change-Id: Idc7bca820e472e281a3024a5b610fd55606cf428