Fix for bug 6716343. Use correct offset for mPos reset. DO NOT MERGE
The mPos field in the MeasuredText object is relative to the start of
the text (mTextStart), but the pos passed in by the caller of the
setPos() method is relative to the character sequence. When spans
overlap break boundaries and the paragraph doesn't start at 0, the
result is an out of bounds error. This fix uses the correct offset.

Change-Id: I64ef06df0eb06f75aedd25de97e9f347eeb52979