Fix camera disambiguation in secure keyguard
When there are multiple activities that respond to MediaStore.INTENT_ACTION_STILL_IMAGE_CAMERA_SECURE
we need to show the disambiguation dialog to the user.  However the disambiguation "dialog"
is actully an activity themed as a dialog.  Hence, we can't show it in the secure keyguard.

This works around the issue by prompting the user for their credentials directly when the
intent needs disambiguation.  This will take them out of keyguard and prompt them for
which activity they want to use.

We'll provide a more robust solution in a future release.

Fixes bug 7109816

Change-Id: I94e643d3cb503e1ce6de24c82400b4d5fcbb9d95